change to double (to represent float64)

+++ src/drivers/mount/mount_mavlink.h
@@ -47,7 +47,7 @@
  bool mount_mavlink_init(int sysid, int compid);
  void mount_mavlink_deinit(void);
  void mount_mavlink_configure(int roi_mode, bool man_control);
- void mount_mavlink_point_location(float lat, float lon, float alt);
+ void mount_mavlink_point_location(double lat, double lon, float alt);
  void mount_mavlink_point_manual(float roll, float pitch, float yaw);
